[Frontend] Refactor Font Optimization
