+++ frontend/index.html
@@ -17,7 +17,7 @@
       property="og:description"
       content="가까운 곳에서 철봉 위치를 찾고 운동에 참여하세요! 철봉 맵은 전국 공원의 철봉 위치를 사용자가 직접 등록하고 조회할 수 있는 플랫폼입니다."
     />
-    <meta property="og:image" content="/images/metaimg.png" />
+    <meta property="og:image" content="/images/metaimg.webp" />
 
     <meta property="twitter:card" content="summary_large_image" />
     <meta property="twitter:url" content="https://chulbong-kr.vercel.app" />
@@ -26,10 +26,18 @@
       property="twitter:description"
       content="철봉 맵은 전국 공원의 철봉 위치를 사용자가 직접 등록하고 조회할 수 있는 플랫폼입니다."
     />
-    <meta property="twitter:image" content="/images/metaimg.png" />
+    <meta property="twitter:image" content="/images/metaimg.webp" />
 
     <title>대한민국 철봉 지도</title>
-    <link rel="icon" type="image/png" sizes="60x60" href="/images/logo.png" />
+    <link rel="icon" type="image/webp" sizes="60x60" href="/images/logo.webp" />
+
+    <link
+      rel="preload"
+      href="/fonts/nunito-v16-latin-regular.woff2"
+      as="font"
+      type="font/woff2"
+      crossorigin="anonymous"
+    />
 
     <script
       type="text/javascript"
