+++ frontend/src/components/AddChinupBarForm/AddChinupBarForm.tsx
@@ -1,5 +1,6 @@
 import type { MarkerClusterer } from "@/types/Cluster.types";
-import { Button, CircularProgress } from "@mui/material";
+import Button from "@mui/material/Button";
+import CircularProgress from "@mui/material/CircularProgress";
 import { useState } from "react";
 import setNewMarker from "../../api/markers/setNewMarker";
 import activeMarkerImage from "../../assets/images/cb1.png";
