+++ frontend/src/components/Map/Map.tsx
@@ -6,11 +6,12 @@ import GpsOffIcon from "@mui/icons-material/GpsOff";
 import LoginIcon from "@mui/icons-material/Login";
 import MyLocationIcon from "@mui/icons-material/MyLocation";
 import RemoveIcon from "@mui/icons-material/Remove";
-import { Button, CircularProgress } from "@mui/material";
+import Button from "@mui/material/Button";
+import CircularProgress from "@mui/material/CircularProgress";
 import { Suspense, lazy, useEffect, useRef, useState } from "react";
 import getAllMarker from "../../api/markers/getAllMarker";
-import activeMarkerImage from "../../assets/images/cb1.png";
-import pendingMarkerImage from "../../assets/images/cb2.png";
+import activeMarkerImage from "../../assets/images/cb1.webp";
+import pendingMarkerImage from "../../assets/images/cb2.webp";
 import useMap from "../../hooks/useMap";
 import useModalStore from "../../store/useModalStore";
 import useUploadFormDataStore from "../../store/useUploadFormDataStore";
