+++ frontend/src/components/SignupForm/SignupForm.tsx
@@ -1,4 +1,5 @@
-import { Button, CircularProgress } from "@mui/material";
+import Button from "@mui/material/Button";
+import CircularProgress from "@mui/material/CircularProgress";
 import { useEffect, useState } from "react";
 import signin from "../../api/auth/signin";
 import useInput from "../../hooks/useInput";
