+++ frontend/src/style.css
@@ -39,7 +39,8 @@ li {
   font-style: normal;
   font-weight: 400;
   src: local(""),
-    url("assets/fonts/nunito-v16-latin-regular.woff2") format("woff2");
+    url("/fonts/nunito-v16-latin-regular.woff2") format("woff2");
+  font-display: swap;
 }
 
 #app {
